Add files via upload
Introduction:
Mr. John Hughes would like to understand the following:
The effect of smoking on expenses
The effect of all input variables on expenses
In order to help Mr. John Hughes, understand the prior, I will conduct a simple linear regression model for the former point. Additionally, I will conduct a multivariate regression model for the latter.
The end goal is to present the data in a way that would be easy to understand for Mr. John Hughes. 
Conclusion:
I have generated both a simple and multivariate regression model for Mr. John Hughes. Additionally, I have also explained my findings in a manner that is accurate and backed by evidence. 
The simple regression model is deemed to be accurate as it fully satisfies the permutations of the number of stars, as well as the adjusted R-value.
Additionally, the multivariate regression model can be accurate to an extent. Only some coefficients have stars associated, whilst others do not. However, the adjusted r-squared value in this model is positive which can deem accuracy to an extent.